- Implementando o checkout - Criando Action de Checkout no CartController
