+++ GeekShopping.Web/Views/Cart/Confirmation.cshtml
@@ -1,0 +1,5 @@
+﻿<div class="text-center">
+    <h1 class="text-info pt-4">Congratulations! You are one step closer to geek lifestyle.</h1>
+    <p class="pb-3">We have received your order and your payment will be processed soon!</p>
+    <img src="/images/order.png" style="width: 80%" />
+</div>
